fix carousel for small devices, some refactoring
